tests: separate tests for the ImageElement
